requirements.txt
appnope==0.1.3
asttokens==2.0.7
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
executing==0.9.1
google==3.0.0
google-api-python-client==1.6.7
google-auth==2.9.1
google-auth-oauthlib==0.5.2
gspread==5.4.0
gspread-dataframe==3.3.0
httplib2==0.20.4
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tornado==6.2
traitlets==5.3.0
twilio==7.12.0
uritemplate==3.0.1
urllib3==1.26.11
wcwidth==0.2.5
